refactor: productservice와 productdetailservice 분리
