+++ backend/src/main/java/com/multi/mariage/product/vo/PairingFoodCountsVO.java
@@ -10,7 +10,7 @@ public class PairingFoodCountsVO {
     private int id; // 음식 id
     //    private String imageUrl;    // 음식 이미지 url
     private String name;    // 음식 이름
-    private int count;  // 리뷰 개수
+    private int count;      // 리뷰 개수
 
     private PairingFoodCountsVO(int id, String name, int count) {
         this.id = id;
